pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sdlc-inject"
version = "0.1.0"
description = "CLI tool for injecting SDLC failure patterns into codebases"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "SDLC Training Team" }
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "gitpython>=3.1",
    "jinja2>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[project.scripts]
sdlc-inject = "sdlc_inject.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["sdlc_inject"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
